Show Mark complete option in dialog only when mandatory is 100%
