+++ app/src/main/java/org/eyeseetea/malariacare/layout/dashboard/controllers/AssessModuleController.java
@@ -22,8 +22,10 @@ package org.eyeseetea.malariacare.layout.dashboard.controllers;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
+import android.support.annotation.NonNull;
 import android.util.Log;
 import android.view.View;
+import android.widget.Button;
 
 import org.eyeseetea.malariacare.DashboardActivity;
 import org.eyeseetea.malariacare.R;
@@ -515,4 +517,58 @@ public class AssessModuleController extends ModuleController {
                 R.id.dashboard_details_container);
     }
 
+    public AlertDialog assetsModelDialog(@NonNull final SurveyDB survey) {
+
+        SurveyDialog.Builder builder = SurveyDialog.newBuilder(dashboardActivity, survey);
+
+        final View.OnClickListener editButtonListener = new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                onSurveySelected(survey);
+            }
+        };
+
+        final View.OnClickListener completeButtonListener = new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                onMarkAsCompleted(survey);
+            }
+        };
+
+        final View.OnClickListener deleteButtonListener = new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //this method create a new survey getting the getScheduledDate date of the
+                // oldsurvey, and remove it.
+                SurveyPlanner.getInstance().deleteSurveyAndBuildNext(survey);
+                DashboardActivity.reloadDashboard();
+            }
+        };
+
+        final SurveyDialog surveyDialog =  builder.editButton(editButtonListener)
+                .completeButton(completeButtonListener)
+                .deleteButton(deleteButtonListener)
+                .build();
+
+        getSurveyAnsweredRatioUseCase.execute(survey.getId_survey(),
+                new ISurveyAnsweredRatioCallback() {
+                    @Override
+                    public void nextProgressMessage() {
+
+                    }
+
+                    @Override
+                    public void onComplete(SurveyAnsweredRatio surveyAnsweredRatio) {
+                        boolean isCompulsoryCompleted = surveyAnsweredRatio.isCompulsoryCompleted();
+                        Button button = surveyDialog.getMarkCompleteButton();
+
+                        if(isCompulsoryCompleted)
+                            button.setVisibility(View.VISIBLE);
+                        else
+                            button.setVisibility(View.GONE);
+                    }
+                });
+
+        return surveyDialog;
+    }
 }
