+++ app/src/main/java/org/eyeseetea/malariacare/layout/dashboard/controllers/DashboardController.java
@@ -22,7 +22,6 @@ package org.eyeseetea.malariacare.layout.dashboard.controllers;
 import android.app.AlertDialog;
 import android.content.res.Resources;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
@@ -39,12 +38,10 @@ import org.eyeseetea.malariacare.data.database.model.ProgramDB;
 import org.eyeseetea.malariacare.data.database.model.SurveyDB;
 import org.eyeseetea.malariacare.data.database.model.SurveyScheduleDB;
 import org.eyeseetea.malariacare.data.database.utils.planning.ScheduleListener;
-import org.eyeseetea.malariacare.data.database.utils.planning.SurveyPlanner;
 import org.eyeseetea.malariacare.layout.dashboard.config.DashboardOrientation;
 import org.eyeseetea.malariacare.layout.dashboard.config.DashboardSettings;
 import org.eyeseetea.malariacare.utils.AUtils;
 import org.eyeseetea.malariacare.views.CustomTextView;
-import org.eyeseetea.malariacare.views.SurveyDialog;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -316,42 +313,8 @@ public class DashboardController {
      * @param survey
      */
     public void onAssetsSelected(SurveyDB survey) {
-        assetsModelDialog(survey);
-    }
-
-
-    public AlertDialog assetsModelDialog(@NonNull final SurveyDB survey) {
-
-        SurveyDialog.Builder builder = SurveyDialog.newBuilder(dashboardActivity, survey);
-
-        final View.OnClickListener editButtonListener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onSurveySelected(survey);
-            }
-        };
-
-        final View.OnClickListener completeButtonListener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onMarkAsCompleted(survey);
-            }
-        };
-
-        final View.OnClickListener deleteButtonListener = new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //this method create a new survey getting the getScheduledDate date of the
-                // oldsurvey, and remove it.
-                SurveyPlanner.getInstance().deleteSurveyAndBuildNext(survey);
-                DashboardActivity.reloadDashboard();
-            }
-        };
-
-        return builder.editButton(editButtonListener)
-                .completeButton(completeButtonListener)
-                .deleteButton(deleteButtonListener)
-                .build();
+        AssessModuleController assessModuleController = (AssessModuleController)getModuleByName(AssessModuleController.getSimpleName());
+        assessModuleController.assetsModelDialog(survey);
     }
 
     /**
